move icon component to common folder
