+++ bigbluebutton-html5/imports/ui/components/common/icon/component.jsx
