+++ bigbluebutton-html5/imports/ui/components/debug-window/styles.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import { smPaddingX } from '/imports/ui/stylesheets/styled-components/general';
 import Resizable from 're-resizable';
-import Icon from '/imports/ui/components/icon/component';
+import Icon from '/imports/ui/components/common/icon/component';
 import Button from '/imports/ui/components/common/button/component';
 
 const ToggleLabel = styled.span`
