+++ bigbluebutton-html5/imports/ui/components/dropdown/list/item/styles.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import Icon from '/imports/ui/components/icon/component';
+import Icon from '/imports/ui/components/common/icon/component';
 import { lineHeightComputed } from '/imports/ui/stylesheets/styled-components/typography';
 import {
   colorGrayDark,
