+++ bigbluebutton-html5/imports/ui/components/nav-bar/styles.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import Icon from '../icon/component';
+import Icon from '/imports/ui/components/common/icon/component';
 import { barsPadding, borderSize } from '/imports/ui/stylesheets/styled-components/general';
 import {
   colorWhite,
