+++ bigbluebutton-html5/imports/ui/components/video-provider/video-list/video-list-item/styles.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import Icon from '/imports/ui/components/icon/component';
+import Icon from '/imports/ui/components/common/icon/component';
 import Button from '/imports/ui/components/common/button/component';
 import {
   colorPrimary,
